move main loop to its own class
